requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
chess==1.11.2
discord.py==2.6.4
python-dotenv==1.1.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
python-chess==1.999
requests==2.32.5
typing-extensions==4.15.0
urllib3==2.5.0
watchdog==6.0.0
yarl==1.22.0
stockfish